remove predefined/default scala.minor.version as this confuses IntelliJ's module library path
If there is a default scala.minor.version of e.g. 2.11 and you enable profile 2.12 in IntelliJ, it will add libraries for 2.11 and 2.12 in library path for modules sdl-kafka/jms/splunk (interestingly not for sdl-core), and compilation in IntelliJ will fail with ClassNotFoundException scala.Product during Macro expansion. This is because a 2.11/2.12 library mismatch - the first in the classpath wins... and scala.Product is not compatible between the two versions.
For compilation with maven this is no problem...