fix conflict between scala-cross-maven-plugin and scala-doc-nolinkwarnings profile
scala-cross-maven-plugin searchs for an active profile starting with "scala-" to replace properties in the pom. Unfortunately during deploy scala-doc-nolinkwarnings was enabled and choosen. As a result no expressions ${scala.minor.version} were replaced in the pom.
This fix removes scala-doc-nolinkwarnings completely and configures execution for creation of scala-doc directly in the release-sonatype profile, including "-no-link-warnings" argument.