502 - changeDateColumnName replaced with incrementalOutputExpr
Previously the changeDateColumnName was set by the current time of the SDLB execution engine.
The incrementalOutputExpr is based on the incoming data. But it requires for the column
to be present in the output and it causes the system to read the received dataframe twice.
Once to determine the new high-water-mark and once for the actual writing process.